+++ amp_funcs.py
@@ -3,7 +3,7 @@ import pandas as pd
 from Subjects import EegRecordSubject
 from Subjects import Participant
 
-def get_amp_diff_data(path: str, channels, events_to_check, tmin: int, tmax: int, max_files = 0):
+def get_amp_diff_data(path: str, channels, events_to_check, tmin: float, tmax: float, max_files = 0):
     df = pd.read_excel("participants.xlsx")
     if max_files == 0:
         max_files = len(df)
